[cffi] add "gruvi" prefix to module names
A reminder why cffi module names are hard coded: we #include source
files in ffi.verify(). These source files are not available at runtime,
so ffi.verify() cannot create the version that goes into the default
module name. The alternative would be either to include the full text of
these sources into the Python file, or to ship the C sources separately
as libraries. Both options are worse in my view.