[futures] make threads in thread pool daemonic
The main Python program will wait for all non-daemonic threads to exit before
exiting itself. Since the user did not explicitly start the threads in a thread
pool, it is easy to forget about them, making your program unable to exit. So
according to the principle of least surprise, it's better to make these threads
daemonic.

The only drawback is that daemonic threads are killed abruptly when the main
program exits. So you should not to exit() if there's work going on in a thread
pool that is not crash-safe. This is pretty obvious, though.