[ssl] do_handshake_on_connect improvements
This changes the behavior of create_connection() so that in case of
do_handshake_on_connect=True (the default), it will wait for the SSL
handshake to complete.  This is useful in various applications that want
to access the SSL state after a connection has been established but
before any application-level data has been read.