[stream] more useful EOF behavior
This will now return EOF (i.e. read() returns b'') on a stream protocol
when the connection was lost and there is an error. On SSL it is happens
that the stream ends up in an error state when the remote peer sends a
close_notify and immediately closes the connection. This is allowed by
the RFC, and it will result in our write() of the close_notify
acknowledge failing with EPIPE.