fix raising saved exceptions
From the patch:

Support for re-raising stored exceptions. In our callback based code it
is very common that we store an exception in a callback as an instance
attribute, and then raise that from the consumer API. This is
problematic on Python3 because the traceback in Py3K is attached to the
exception itself (as the __traceback__ attribute). The impact is that i)
the traceback will also be stored and never freed, and ii) if the
exception is re-raised multiple times a new traceback will be appended
for every raise.

The solution that we're using is is to have a saved_exc() call that
returns a new copy of the exception on Python3.