OSX Mavericks sslcompat no fudge
On previous versions of OSX a fudge offset of 8 bytes was needed for the SSLv3
offset in the SSL context. On Mavericks this doesn't appear to be necessary
anymore. On Mavericks, "gcc" is now a symlink to clang, so the original problem
was very likely a gcc vs clang issue.