Allow _sslcompat to be absent on 2.x on Windows.
Building _sslcompat on Python 2.x on Windows is too tricky. There is no
system provided version of OpenSSL, and the official Python binaries use
a static version that is embedded during the build process.

The only fool proof way to get _sslcompat on Python 2.x on Windows is to
build your own OpenSSL *and* Python.