process: fix race condition in _on_child_exit()
The _on_child_exit handler would unconditionally close the stdio handles,
which is problematic as the handles could still have data in them.

The fix is to not to close the stdio handles when the process exits, and to
add a .close() method that needs to be called explicitly by the user to do
the cleanup.