transports: improve error handling
No longer set an error on EOF. An EOF at the transport level is not
always an error. Python asyncio made the same choice: connection_lost()
will have exc=None on a normal connection close.

An EOF can still be an error at the protocol level of course, if an EOF
happens in violation of a protocol. This is protocol dependent and the
error should be raised there.