hub: change @switchpoint to use a closure
When @switchpoint was first developed, Sphinx would not correctly
document the signatures of functions wrapped by a decorator. As a
workaround, Gruvi implemented a rather complicated `util.wrap()`
function that generated a wrapper function on the fly with the same
signature and name as the wrapped function, so that Sphinx would
correctly document it.

Sphinx has been fixed for a while now, by honoring the `__wrapped__`
attribute that `functools.wraps` creates. So we can use that now to wrap
a closure, and remove a bunch of code.

This commit is the first one in a sequence to remove `util.wrap`.