core: refactor flow control
Move read-side flow control away from the Protocol base class and into
the specific subclasses. This removes a pretty thin interface between
Protocol and its subclasses. Its just as easy to call pause/resume
reading directly when parsing the protocol.

Move write-site flow control into the Transport base class. Protocols
that write to a transport should make sure Transport._can_write is
set, and if not wait for it.

With these changes, Protocol is now only an interface definition, and it
is no longer needed to do supercalls from protocol callbacks.

This also does away with StreamReader and StreamWriter. The
ReadWriteStream class was renamed to Stream and now handles all of this,
with the help of a simple StreamBuffer utility class. This removes the
duplication where we'd have to use different classes based on the
read/write mode. I prefer a single class with a "mode" argument, just
like files. This also how Transport works.