examples: fix encoding issue curl.py
A test was failing because we're reading the body in 4k chunks, and a
chunk happened to contain a partial UTF-8 mulit-byte encoded codepoint.

For text bodies, the correct way would be to decode the body using
TextIOWrapper() and then encode using the encoding on sys.stdout. For
binary, it would be to dump the binary to the "raw" standard output
(sys.stdout.buffer). And on Python 2.7 air need to be taken that in case
output is redirected, there will be no output encoding.

However since this is an example just keep it simple, and write the
response unencoded to the terminal witout. This will work on both Python
2 and Python 3, and whether output is redirected or not. The small-ish
issue is when a text body has a different encoding than the terminal but
this usually is not a problem (and the real curl doesn't appear to solve
this either).