ssl: limit support for Python 2.7.8 and below
The sslcompat modules contained two backports: the async SSL support
from 3.5, and backports of other functions that were added over time to
increase security and functinality (e.g. ALPN and ECDH support).

Since 2.7.9 and higher now has a recent _ssl module, I am removing all
backports except the async SSL support. This simplifies the code quite a
bit. Version 2.7.8 and below are still supported, but only with the
feature set that came out of the box in the stdlib.

Another reason to do this is that the added security features didn't
make sense in 2.7.8 anyway, as 2.7.8 didn't do certificate validation
which completely undermines any additional security capabilities on top.

In summary, to get more SSL capabilities, use a recent 2.7.x or 3.y
release.