ssl: verify certs by default
This add a function sslcompat.create_default_context() that forwards to
ssl.create_default_context() on Python versions that have it, and
otherwise emulates it. This function returns an SSLContext with the best
balance between interoperability and security for the capabilities of
the _ssl module of the specific Python version.

This new function is now the default context for create_connection() and
create_server(), thereby enforcing certificate validation on any Python
version that supports it (i.e. >= 2.7.9).

All tests now validate certificates also.