hub: optimize @switchpoint
This is a deep change that changes the logic of our safety check around
switchpoints. There's two checks that we want to enforce:

1. No switchpoints are called from loop callbacks. This is mostly for
   keeping a sane call graph where we don't switch out from under libuv
   (even though this would theoretically work).
2. Only the hub can cause the actual switch to happen (in the outer
   loop, outside the libuv event loop). This ensures the hub is the
   parent for all fibers, and gives us a sane "hub and spoke" switching
   model.

Previously these checks were enforced by a wrapped function returned by
the @switchpoint decorator. Profiling turned out however that these
checks were adding a 10% overhead on the HTTP benchmarks, which is a bit
much. In the new model, the checks are only performed when switch() is
actually called. So you don't get an early error, but you will get the
error eventually if a switch does happen.